requirements.txt
numpy>=1.24,<2.0
opencv-python>=4.8,<4.11
mediapipe>=0.10.8,<0.11
matplotlib>=3.7,<3.10
Pillow>=10.0,<11
